requirements/base.txt
Django==1.7
PyYAML==3.11
argparse==1.2.1
django-fontawesome==0.1
django-grappelli==2.6.1
wsgiref==0.1.2
